Fix navigation behaviour starting to level destination from deeper backstack (#14)

+++ app/src/main/java/at/aau/iteractivesystems/library/ui/bookdetail/DetailViewModel.kt
@@ -4,8 +4,10 @@ import androidx.lifecycle.*
 import at.aau.iteractivesystems.library.Environment
 import at.aau.iteractivesystems.library.repository.books.BookRepository
 import at.aau.iteractivesystems.library.repository.books.BorrowedBooksRepository
+import at.aau.iteractivesystems.library.repository.user.UserRepository
 import at.aau.iteractivesystems.library.ui.adapter.Content
 import at.aau.iteractivesystems.library.ui.utils.AndroidString
+import at.aau.iteractivesystems.library.ui.utils.SingleLiveEvent
 import at.aau.iteractivesystems.library.ui.utils.ViewState
 import kotlinx.coroutines.launch
 
@@ -13,6 +15,7 @@ class DetailViewModel(
     private val bookId: String,
     private val bookRepository: BookRepository,
     private val borrowedBooksRepository: BorrowedBooksRepository,
+    private val userRepository: UserRepository,
 ) : ViewModel() {
 
     private val _state: MutableLiveData<ViewState<List<Content>>> =
@@ -24,17 +27,28 @@ class DetailViewModel(
     val borrowed: LiveData<Boolean>
         get() = _borrowed
 
+    private val _navigateToLogin = SingleLiveEvent<Unit>()
+    val navigateToLogin: LiveData<Unit>
+        get() = _navigateToLogin
+
     init {
         setupContent()
     }
 
     fun toggleBorrowed() {
-        if (borrowedBooksRepository.contains(bookId)) {
-            borrowedBooksRepository.remove(bookId)
-            _borrowed.postValue(false)
-        } else {
-            borrowedBooksRepository.add(bookId)
-            _borrowed.postValue(true)
+        viewModelScope.launch {
+            if (userRepository.getUser() == null) {
+                _navigateToLogin.postValue(Unit)
+                return@launch
+            }
+
+            if (borrowedBooksRepository.contains(bookId)) {
+                borrowedBooksRepository.remove(bookId)
+                _borrowed.postValue(false)
+            } else {
+                borrowedBooksRepository.add(bookId)
+                _borrowed.postValue(true)
+            }
         }
     }
 
@@ -78,7 +92,8 @@ class DetailViewModel(
                 DetailViewModel(
                     bookId = bookId,
                     bookRepository = environment.bookRepository,
-                    borrowedBooksRepository = environment.borrowedBooksRepository
+                    borrowedBooksRepository = environment.borrowedBooksRepository,
+                    userRepository = environment.userRepository,
                 ) as T
             } else throw IllegalArgumentException("Unknown ViewModel class.")
         }
